Se agregaron elementos y se testetaron las posiciones

+++ CalculadoraIMC/ViewController.swift
@@ -12,9 +12,9 @@ class ViewController: UIViewController {
     @IBOutlet weak var lblPeso: UILabel!
     @IBOutlet weak var lblAltura: UILabel!
     var cerebroCalculadora = CerebroCalculadora()
-    var peso:float = 75.0
-    var altura:float = 1.6
-    var imc:float = 0.0
+    var peso:Double = 75.0
+    var altura:Double = 1.6
+    var imc:Double = 0.0
     
 
     override func viewDidLoad() {
@@ -44,7 +44,7 @@ class ViewController: UIViewController {
     
     @IBAction func btnCalcular(_ sender: UIButton) {
         print("calcular")
-        cerebroCalculadora.calcularIMC(peso: peso, altura: altura)
+        cerebroCalculadora.calcularIMC(peso: Float(peso), altura: Float(altura))
         performSegue(withIdentifier: "calcularSegue", sender: self)
